requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
bs4
pypdf
pymupdf
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
beautifulsoup4
streamlit
langchain_groq
langchain_core
fastapi
langserve[all]
arxiv
wikipedia
duckduckgo-search
SQLAlchemy
mysql-connector-python
langchain-pinecone
sqlitecloud
streamlit-option-menu
redis
streamlit-lottie